fix: propagate fallback model through gateway resolution
