+++ tests/gateway/test_auth_fallback.py
@@ -22,15 +22,15 @@ class TestResolveRuntimeAgentKwargsAuthFallback:
 
         monkeypatch.setattr("gateway.run._hermes_home", tmp_path)
 
-        call_count = {"n": 0}
+        calls = []
 
         def _mock_resolve(**kwargs):
-            call_count["n"] += 1
+            calls.append(kwargs)
             # First call = primary path (gateway reads model.provider from
             # config.yaml internally; we simulate the auth failure here).
             # Second call = fallback path with explicit_api_key + explicit_base_url
             # supplied by gateway from fallback_model config.
-            if call_count["n"] == 1:
+            if len(calls) == 1:
                 raise AuthError("Codex token refresh failed with status 401")
             return {
                 "api_key": "fallback-key",
@@ -51,7 +51,8 @@ class TestResolveRuntimeAgentKwargsAuthFallback:
 
         assert result["provider"] == "openrouter"
         assert result["api_key"] == "fallback-key"
-        # Should have been called at least twice (primary + fallback)
-        assert call_count["n"] >= 2
+        assert len(calls) >= 2
+        assert calls[1]["requested"] == "openrouter"
+        assert calls[1]["target_model"] == "meta-llama/llama-4-maverick"
